backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
spacy
pdfplumber
pytesseract
Pillow
PyMuPDF
